[🔄 Refactor] 코드 구조 개선 (#168)
* fix: select에서 연,월 선택시 state 값이 바뀌도록 수정

* refactor: 관리자 급여 정정신청 내역 페이지 급여에 , 구분자 추가

* refactor: 관리자 급여 페이지 데이터 관련 내용 파일 분리
